+++ src/rpc/iterative_query.rs
@@ -21,10 +21,10 @@ pub(crate) struct IterativeQuery {
     pub request: RequestSpecific,
     closest: ClosestNodes,
     responders: ClosestNodes,
-    inflight_requests: Vec<u16>,
+    inflight_requests: Vec<u32>,
     visited: HashSet<SocketAddrV4>,
     responses: Vec<Response>,
-    public_address_votes: HashMap<SocketAddrV4, u16>,
+    public_address_votes: HashMap<SocketAddrV4, u32>,
 }
 
 #[derive(Debug)]
@@ -93,7 +93,7 @@ impl IterativeQuery {
     }
 
     pub fn best_address(&self) -> Option<SocketAddrV4> {
-        let mut max = 0_u16;
+        let mut max = 0_u32;
         let mut best_addr = None;
 
         for (addr, count) in self.public_address_votes.iter() {
@@ -146,7 +146,7 @@ impl IterativeQuery {
     }
 
     /// Return true if a response (by transaction_id) is expected by this query.
-    pub fn inflight(&self, tid: u16) -> bool {
+    pub fn inflight(&self, tid: u32) -> bool {
         self.inflight_requests.contains(&tid)
     }
